Move group to file
